requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.4
imbalanced-learn>=0.12
matplotlib>=3.7
seaborn>=0.13
PyYAML>=6.0
pytest>=8.0
